add manager files, change providers, remove rotor
