--- pkg/certificate/providers/certmanager/errors.go
@@ -1,7 +1,0 @@
-package certmanager
-
-import (
-	"errors"
-)
-
-var errCertNotFound = errors.New("failed to find cert")
