+++ pkg/certificate/providers/tresor/types.go
@@ -3,13 +3,9 @@ package tresor
 
 import (
 	"math/big"
-	"sync"
-	"time"
 
 	"github.com/openservicemesh/osm/pkg/certificate"
-	"github.com/openservicemesh/osm/pkg/configurator"
 	"github.com/openservicemesh/osm/pkg/logger"
-	"github.com/openservicemesh/osm/pkg/messaging"
 )
 
 const (
@@ -26,20 +22,11 @@ var (
 )
 
 // CertManager implements certificate.Manager
-type CertManager struct {
+type Provider struct {
 	// The Certificate Authority root certificate to be used by this certificate manager
 	ca *certificate.Certificate
 
-	// Cache for all the certificates issued
-	// Types: map[certificate.CommonName]*certificate.Certificate
-	cache sync.Map
-
 	certificatesOrganization string
 
-	cfg configurator.Configurator
-
-	serviceCertValidityDuration time.Duration
-	keySize                     int
-
-	msgBroker *messaging.Broker
+	keySize int
 }
