+++ pkg/certificate/providers/vault/certificate_manager.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/openservicemesh/osm/pkg/certificate"
 	"github.com/openservicemesh/osm/pkg/certificate/pem"
-	"github.com/openservicemesh/osm/pkg/certificate/rotor"
 	"github.com/openservicemesh/osm/pkg/constants"
 	"github.com/openservicemesh/osm/pkg/errcode"
 	"github.com/openservicemesh/osm/pkg/logger"
@@ -30,33 +29,35 @@ const (
 	decade                             = 8765 * time.Hour
 )
 
-//todo(schristoff): NewManagement, VaultManager & Certificate.Manager
-// NewCertManager implements certificate.Manager and wraps a Hashi Vault with methods to allow easy certificate issuance.
-func NewVaultManager(vaultAddr string, token string, role string) (*CertManager, error) {
-	v := &VaultManager{
+// NewProvider implements certificate.Manager and wraps a Hashi Vault with methods to allow easy certificate issuance.
+func NewProvider(vaultAddr string, token string, role string) (*Provider, error) {
+	p := &Provider{
 		role: vaultRole(role),
 	}
-
-	c := certificate.NewManager(v)
 
 	config := api.DefaultConfig()
 	config.Address = vaultAddr
 
 	var err error
-	if v.client, err = api.NewClient(config); err != nil {
+	if p.client, err = api.NewClient(config); err != nil {
 		return nil, errors.Errorf("Error creating Vault CertManager without TLS at %s", vaultAddr)
 	}
 
 	log.Info().Msgf("Created Vault CertManager, with role=%q at %v", role, vaultAddr)
 
-	v.client.SetToken(token)
+	p.client.SetToken(token)
 
-	issuingCA, serialNumber, err := c.getIssuingCA(c.IssueCertificate)
+	return p, nil
+}
+
+func (p *Provider) IssueCertificate(cn certificate.CommonName, validityPeriod time.Duration) (*certificate.Certificate, error) {
+
+	issuingCA, serialNumber, err := p.GetIssuingCA(p.IssueCertificate)
 	if err != nil {
 		return nil, err
 	}
 
-	c.ca = &certificate.Certificate{
+	ca := &certificate.Certificate{
 		CommonName:   constants.CertificationAuthorityCommonName,
 		SerialNumber: serialNumber,
 		Expiration:   time.Now().Add(decade),
@@ -65,13 +66,9 @@ func NewVaultManager(vaultAddr string, token string, role string) (*CertManager,
 	}
 
 	// Instantiating a new certificate rotation mechanism will start a goroutine for certificate rotation.
-	rotor.New(c).Start(checkCertificateExpirationInterval)
+	certificate.New(c).Start(checkCertificateExpirationInterval)
 
-	return c, nil
-}
-
-func (v *VaultManager) IssueCertificate(cn certificate.CommonName, validityPeriod time.Duration) (*certificate.Certificate, error) {
-	secret, err := cm.client.Logical().Write(getIssueURL(cm.role).String(), getIssuanceData(cn, validityPeriod))
+	secret, err := p.client.Logical().Delete().v.Logical().Write(getIssueURL(cm.role).String(), getIssuanceData(cn, validityPeriod))
 	if err != nil {
 		// TODO(#3962): metric might not be scraped before process restart resulting from this error
 		log.Error().Err(err).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrIssuingCert)).
@@ -91,7 +88,7 @@ func (v *VaultManager) IssueCertificate(cn certificate.CommonName, validityPerio
 }
 
 //todo(schristoff): is this needed?
-func (cm *CertManager) getIssuingCA(issue func(certificate.CommonName, time.Duration) (*certificate.Certificate, error)) ([]byte, certificate.SerialNumber, error) {
+func (p *Provider) getIssuingCA(issue func(certificate.CommonName, time.Duration) (*certificate.Certificate, error)) ([]byte, certificate.SerialNumber, error) {
 	// Create a temp certificate to determine the public part of the issuing CA
 	cert, err := issue("localhost", decade)
 	if err != nil {
@@ -101,20 +98,19 @@ func (cm *CertManager) getIssuingCA(issue func(certificate.CommonName, time.Dura
 	issuingCA := cert.GetIssuingCA()
 
 	// We are not going to need this certificate - remove it
-	cm.ReleaseCertificate(cert.GetCommonName())
+	p.ReleaseCertificate(cert.GetCommonName())
 
 	return issuingCA, cert.GetSerialNumber(), err
 }
 
 // ReleaseCertificate is called when a cert will no longer be needed and should be removed from the system.
-func (v *VaultManager) ReleaseCertificate(cn certificate.CommonName) {
+func (p *Provider) ReleaseCertificate(cn certificate.CommonName) {
 	// TODO(draychev): implement Hashicorp Vault delete-cert API here: https://github.com/openservicemesh/osm/issues/2068
-	v.client.Logical().Delete()
-	cm.deleteFromCache(cn)
-
-}
-
-//todo(schristoff): is this needed? could just put in line.
-func (cm *CertManager) deleteFromCache(cn certificate.CommonName) {
-	cm.cache.Delete(cn)
+	secret, err := p.client.Logical().Delete().v.Logical().Write(getIssueURL(cm.role).String(), getIssuanceData(cn, validityPeriod))
+	if err != nil {
+		// TODO(#3962): metric might not be scraped before process restart resulting from this error
+		log.Error().Err(err).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrIssuingCert)).
+			Msgf("Error issuing new certificate for CN=%s", cn)
+		return nil, err
+	}
 }
