--- pkg/certificate/rotor/suite_test.go
@@ -1,13 +1,0 @@
-package rotor
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestRotor(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Test Suite")
-}
